requirements.txt
fastapi = 0.101.1
SQLAlchemy = 2.0.20
passlib = 1.7.4
PyJWT = 2.8.0
python-jose = 3.3.0
pydantic = 2.3.0
pydantic_core = 2.6.3
PyMySQL = 1.1.0
